% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{effectSize}
\alias{effectSize}
\title{Calculate Cohen's d (effect size)}
\usage{
effectSize(lowestMean, lowestSD, highestMean, highestSD)
}
\arguments{
\item{lowestMean}{Mean in lowest group}

\item{highestMean}{Mean in highest group}

\item{highestSD}{SD in highest group}

\item{LowestSD}{SD in lowest group}
}
\value{
Effect size estimate
}
\description{
Helper method to calculate effect size using Cohen's d formula.
}
