% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotVarAndQuintile}
\alias{plotVarAndQuintile}
\title{Plot variables within quintiles of another variable}
\usage{
plotVarAndQuintile(dataInput, exposure, outcome, writePDF = FALSE)
}
\arguments{
\item{dataInput}{Dataset to use}

\item{exposure}{Name of exposure variable}

\item{outcome}{Name of outcome variable}

\item{writePDF}{Logical parameter. If TRUE, writes pdf file with output.}
}
\description{
Plot variables within quintiles of another variable
}
