% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pValueHTML_cell}
\alias{pValueHTML_cell}
\title{Format p-value}
\usage{
pValueHTML_cell(pVal, cut)
}
\arguments{
\item{pVal}{p-value}

\item{cut}{Smallest p-value to be shown}
}
\value{
Cell for HTML summary table with formatted p-values.
}
\description{
Helper method to show formatted p-value in HTML table
}
