% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary4tableHTML_simple}
\alias{summary4tableHTML_simple}
\title{Format mean and SD}
\usage{
summary4tableHTML_simple(col, numDigits)
}
\arguments{
\item{col}{A numeric vector}

\item{numDigits}{Number of decimal places for output.}
}
\value{
Cell for HTML summary table giving mean and SD of vector to a specified number of decimal places.
}
\description{
Helper method to show formatted mean and SD to specified number of decimal places for use in table cell.
}
